Finally fixed editing of the widget and also moved widget to a separate file
